EntityFader and Tutorial tweaks
Set the x, y values for EntityFader objects so they appear properly and not in the top left of the screen. Added increment/decrements to the tween step for EntityFader. Removed test02.fadeOut() from Tutorial and changed a pauseCounter++ to += 2 in order to display the third text field and not crash. This last change is a temporary work around.