Fixed third tutorial text not displaying, various minor fixes and additions
Due to a syntax error in the original code, the Tutorial object step event had to be changed to reflect this functionality. Removed the if check and changed it to be a simple "if(true)" as it should always trigger.

For various fixes, added a self-destruct for HeartBeatFlat objects in their step when alpha reaches 0 to match Heartbeat functionality. Added extra functions to cheater class for testing. Changed "image.color" to "image_blend" to match GMS2 colors.